Sacado:  Move LayoutContiguous to Sacado namespace.
