+++ packages/sacado/test/UnitTests/Fad_KokkosAtomicTests_Cuda_Hierarchical_DFad.cpp
@@ -17,8 +17,8 @@
 
 #include "Fad_KokkosAtomicTests.hpp"
 
-typedef Kokkos::LayoutContiguous<Kokkos::LayoutLeft,32> LeftContiguous32;
-typedef Kokkos::LayoutContiguous<Kokkos::LayoutRight,32> RightContiguous32;
+typedef Sacado::LayoutContiguous<Kokkos::LayoutLeft,32> LeftContiguous32;
+typedef Sacado::LayoutContiguous<Kokkos::LayoutRight,32> RightContiguous32;
 #undef VIEW_FAD_TESTS_FD
 #define VIEW_FAD_TESTS_FD( F, D )                                       \
   VIEW_FAD_TESTS_FLD( F, LeftContiguous32, D )                          \
