+++ packages/sacado/test/UnitTests/Fad_KokkosTests_Cuda_Hierarchical_SFad.cpp
@@ -18,8 +18,8 @@
 
 #include "Fad_KokkosTests.hpp"
 
-typedef Kokkos::LayoutContiguous<Kokkos::LayoutLeft,32> LeftContiguous32;
-typedef Kokkos::LayoutContiguous<Kokkos::LayoutRight,32> RightContiguous32;
+typedef Sacado::LayoutContiguous<Kokkos::LayoutLeft,32> LeftContiguous32;
+typedef Sacado::LayoutContiguous<Kokkos::LayoutRight,32> RightContiguous32;
 #undef VIEW_FAD_TESTS_FDC
 #define VIEW_FAD_TESTS_FDC( F, D )                                      \
   VIEW_FAD_TESTS_FLD( F, LeftContiguous32, D )                          \
