+++ packages/sacado/test/UnitTests/Fad_KokkosTests_HIP.cpp
@@ -34,7 +34,7 @@ TEUCHOS_UNIT_TEST(Kokkos_View_Fad, SFadHipAligned)
   const int Stride = 32;
   const int LocalDim = 2;
   typedef Sacado::Fad::SFad<double,StaticDim> FadType;
-  typedef Kokkos::LayoutContiguous<Kokkos::LayoutLeft,Stride> Layout;
+  typedef Sacado::LayoutContiguous<Kokkos::LayoutLeft,Stride> Layout;
   typedef Kokkos::HIP Device;
   typedef Kokkos::View<FadType*,Layout,Device> ViewType;
 
@@ -59,7 +59,7 @@ TEUCHOS_UNIT_TEST(Kokkos_View_Fad, SFadHipNotAligned)
   const int Stride = 32;
   const int LocalDim = 0;
   typedef Sacado::Fad::SFad<double,StaticDim> FadType;
-  typedef Kokkos::LayoutContiguous<Kokkos::LayoutLeft,Stride> Layout;
+  typedef Sacado::LayoutContiguous<Kokkos::LayoutLeft,Stride> Layout;
   typedef Kokkos::HIP Device;
   typedef Kokkos::View<FadType*,Layout,Device> ViewType;
