+++ packages/sacado/test/performance/advection_const_basis/advection_hierarchical_dfad.cpp
@@ -107,7 +107,7 @@ double time_dfad_hierarchical_team(int ncells, int num_basis, int num_points,
   typedef Sacado::Fad::DFad<double> FadType;
 
   typedef typename ExecSpace::array_layout DefaultLayout;
-  typedef Kokkos::LayoutContiguous<DefaultLayout> ContLayout;
+  typedef Sacado::LayoutContiguous<DefaultLayout> ContLayout;
   typedef Kokkos::View<double****,ExecSpace> t_4DView_d;
   typedef Kokkos::View<double***,ExecSpace> t_3DView_d;
   typedef Kokkos::View<FadType***,ContLayout,ExecSpace> t_3DView;
@@ -164,7 +164,7 @@ double time_dfad_hierarchical_team_scratch(
   typedef Sacado::Fad::DFad<double> FadType;
 
   typedef typename ExecSpace::array_layout DefaultLayout;
-  typedef Kokkos::LayoutContiguous<DefaultLayout> ContLayout;
+  typedef Sacado::LayoutContiguous<DefaultLayout> ContLayout;
   typedef Kokkos::View<double****,ExecSpace> t_4DView_d;
   typedef Kokkos::View<double***,ExecSpace> t_3DView_d;
   typedef Kokkos::View<FadType***,ContLayout,ExecSpace> t_3DView;
